Move coordinators together into a package of their own
