+++ reasoner/resolution/resolver/coordinator/Coordinator.java
@@ -13,9 +13,10 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
  */
 
-package com.vaticle.typedb.core.reasoner.resolution.resolver;
+package com.vaticle.typedb.core.reasoner.resolution.resolver.coordinator;
 
 import com.vaticle.typedb.core.concept.ConceptManager;
 import com.vaticle.typedb.core.concept.answer.ConceptMap;
